CustomLocationManager
