+++ WeatherApp/WeatherApp/Model/LocationManager.swift
@@ -1,0 +1,15 @@
+//
+//  LocationManager.swift
+//  WeatherApp
+//
+//  Created by xcode on 12.11.16.
+//  Copyright © 2016 VSU. All rights reserved.
+//
+
+import Foundation
+import CoreLocation
+
+class LocationManager{
+    let locManager = CLLocationManager()
+    
+}
